[TF][MAPS] support geo: intents
This patch adds the ability for Telegram to parse locations from send
intents containing a geo:<lat>,<lon>,<zoom> string.

This is done by various map applications, including the popular OSMAnd
and Orux maps. This patch enables the user to share the position of a
POI or destination with Telegram and have it send as a location.

(cherry picked from commit 0d7de43efe0469ec8f93a07072900bfc1ac47533)